requirements.txt
flask
flask-cors
ccxt
python-dotenv
pandas
numpy
ta
requests